pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aoitalk"
version = "0.1.0"
description = "AI-powered voice conversation agent with Discord integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AoiTalk Team"}
]

# Core dependencies (必須依存 - 会社用エージェントでも使用)
dependencies = [
    # LLM & AI
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "openai-agents>=0.0.16",
    "mcp[cli]>=1.0.0",
    
    # Web Interface (FastAPI)
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    
    # Memory Management (PostgreSQL)
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
    "alembic>=1.12.0",
    "sentence-transformers>=2.2.0",
    "pydantic>=2.0.0",
    "bcrypt>=4.0.0",
    "PyJWT>=2.8.0",
    "itsdangerous>=2.1.0",
    "tiktoken>=0.5.0",
    "psycopg2-binary>=2.9.0",
    "mem0ai>=0.1.0",
    
    # RAG (Qdrant + LlamaIndex + BGE)
    "qdrant-client>=1.16.0,<1.17.0",
    "llama-index>=0.10.0",
    "llama-index-vector-stores-qdrant>=0.1.0",
    "llama-index-embeddings-huggingface>=0.1.0",
    "llama-index-readers-file>=0.1.0",
    "FlagEmbedding>=1.2.0",
    "pypdf>=3.0.0",
    
    # Repository Maps
    "grep-ast>=0.3.0",
    "tree-sitter-language-pack>=0.1.0",
    "diskcache>=5.6",
    "tqdm>=4.65",
    
    # Office File Reading
    "markitdown[docx,pptx,xlsx,pdf]>=0.0.1",
    
    # Video Streaming
    "yt-dlp>=2024.0.0",
    "ffmpeg-python>=0.2.0",
    
    # Graph Analysis
    "python-louvain>=0.16",
    "networkx>=2.8.0",
    "faiss-cpu>=1.7.0",
    
    # Core Utilities
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "httpx>=0.24.0",
    "nest-asyncio>=1.5.0",
    "psutil>=5.9.0",
    "scipy>=1.10.0",
    
    # Spotify (Optional but in core)
    "spotipy>=2.24.0",
]

[project.optional-dependencies]
# Audio processing dependencies (音声入出力・TTS・ASR・再生・録音)
audio = [
    # Basic Audio
    "voicevox-client",
    "sounddevice>=0.4.6",
    "pyaudio",
    "pydub>=0.25.1",
    "pygame>=2.0.0",
    
    # TTS & ASR
    "azure-cognitiveservices-speech>=1.31.0",
    "gTTS>=2.3.2",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "openai-whisper>=20230918",
    "nemo-toolkit[asr]>=1.20.0",
    
    # Discord Voice
    "discord.py>=2.0.0",
    "discord-ext-voice-recv",
    "PyNaCl>=1.5.0",
]

# Windows-specific dependencies (Windows専用)
windows = [
    "pyvcroid2 @ git+https://github.com/Nkyoku/pyvcroid2.git",
    "pythonnet>=3.0.0",
    "pywin32",
]

# Test dependencies (テスト関連)
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.11.0",
    "pytest-xdist>=3.0.0",
    "pytest-timeout>=2.1.0",
]

# Qwen3 TTS dependencies (Qwen TTS関連)
qwen3 = [
    "qwen-tts",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

# pytest設定は pytest.ini で一元管理
